+++ changelog.md
@@ -1,6 +1,6 @@
 ## SpellActivationOverlay Changelog
 
-#### v2.5.3 (2025-06-xx)
+#### v2.5.3 (2025-06-17)
 
 SpellActivationOverlay vs. SpellActivationOverlay
 - The game's spell alert options are now checked at start (Mists of Pandaria)
@@ -15,7 +15,7 @@ Unsupported Spell Alerts (Mists of Pandaria)
 New Spell Alerts
 - Monk's Combo Breaker: Blackout Kick
 - Monk's Combo Breaker: Tiger Palm
-- Rogue's Blindside
+- Rogue's Blindside (Mists of Pandaria)
 
 Internationalization
 - Traditional Chinese (zh-TW) is now fully supported